Reorganize source into microservice packages
