+++ docker-compose.yaml
@@ -2,7 +2,7 @@ version: '3.9'
 services:
   # ergosafe:
   #   image: ergosafe
-  #   command: uvicorn ergosafe.main:app --host 0.0.0.0 --port 8000 --reload
+  #   command: uvicorn ergosafe.api.main:app --host 0.0.0.0 --port 8000 --reload
   #   volumes:
   #     - .:/code
   #   ports:
