+++ ergosafe/influx/influx_conf/query_scores.py
@@ -1,5 +1,10 @@
 from influxdb_client import InfluxDBClient
-from ergosafe.influx_conf.influx_config import INFLUX_BUCKET, INFLUX_URL, INFLUX_TOKEN, INFLUX_ORG
+from ergosafe.influx.influx_conf.influx_config import (
+    INFLUX_BUCKET,
+    INFLUX_URL,
+    INFLUX_TOKEN,
+    INFLUX_ORG,
+)
 
 client = InfluxDBClient(url=INFLUX_URL, token=INFLUX_TOKEN, org=INFLUX_ORG)
